+++ backend/src/User/routes/user.routes.ts
@@ -1,0 +1,10 @@
+import { Router } from 'express';
+import SignUpController from 'User/controllers/SingUpController';
+
+const loginRoutes = Router();
+
+const createUserControler = new SignUpController();
+
+loginRoutes.post('/signup', createUserControler.create);
+
+export default loginRoutes;
